requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
WTForms==3.0.1
email-validator==2.0.0
Werkzeug==2.3.7
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
requests==2.31.0